Remove useGoalSettingsHook
